requirements.txt
Kivy==2.0.0
paho_mqtt==1.5.1
